+++ PokemonApp/Views/Views/PokemonDetailHeaderView.swift
@@ -55,6 +55,30 @@ final class PokemonDetailHeaderView: UIView {
         return label
     }()
     
+    lazy var eggGroups: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    lazy var growthRate: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    lazy var habitat: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    lazy var shape: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
     // MARK: - LifeCycle
     
     override init(frame: CGRect) {
@@ -74,18 +98,22 @@ final class PokemonDetailHeaderView: UIView {
         labelsStackView.addArrangedSubview(baseHappiness)
         labelsStackView.addArrangedSubview(captureRate)
         labelsStackView.addArrangedSubview(generation)
+        labelsStackView.addArrangedSubview(eggGroups)
+        labelsStackView.addArrangedSubview(growthRate)
+        labelsStackView.addArrangedSubview(habitat)
+        labelsStackView.addArrangedSubview(shape)
 
         addSubview(pokemonImage)
         addSubview(labelsStackView)
         
         NSLayoutConstraint.activate([
             
-            pokemonImage.topAnchor.constraint(equalTo: topAnchor),
             pokemonImage.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 8),
-            pokemonImage.bottomAnchor.constraint(equalTo: bottomAnchor),
             
             pokemonImage.heightAnchor.constraint(equalToConstant: 96),
             pokemonImage.widthAnchor.constraint(equalToConstant: 96),
+            
+            pokemonImage.centerYAnchor.constraint(equalTo: centerYAnchor),
             
             labelsStackView.topAnchor.constraint(equalTo: topAnchor, constant: 12),
             labelsStackView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -8),
@@ -99,6 +127,10 @@ final class PokemonDetailHeaderView: UIView {
         let mainPokemonHappiness = "\(L10n.pokemonBaseHappiness)\(pokemonBaseDataModel.baseHappiness)"
         let mainPokemonCaptureRate = "\(L10n.pokemonCaptureRate)\(pokemonBaseDataModel.captureRate)"
         let mainPokemonGeneration = "\(L10n.pokemonGeneration)\(pokemonBaseDataModel.generation)"
+        let mainPokemonEggGroups = "\(L10n.pokemonEggGroups)\(pokemonBaseDataModel.eggGroups)"
+        let mainPokemonGrowthRate = "\(L10n.pokemonGrowthRate)\(pokemonBaseDataModel.growthRate)"
+        let mainPokemonHabitat = "\(L10n.pokemonHabitat)\(pokemonBaseDataModel.habitat)"
+        let mainPokemonShape = "\(L10n.pokemonShape)\(pokemonBaseDataModel.shape)"
         let url = URL(string: pokemonBaseDataModel.pokemonImageUrl)!
         pokemonImage.load(url: url)
         pokemonImage.stopRotation()
@@ -106,6 +138,10 @@ final class PokemonDetailHeaderView: UIView {
         baseHappiness.attributedText = setupAttributedLabels(with: mainPokemonHappiness, and: pokemonBaseDataModel.baseHappiness)
         captureRate.attributedText = setupAttributedLabels(with: mainPokemonCaptureRate, and: pokemonBaseDataModel.captureRate)
         generation.attributedText = setupAttributedLabels(with: mainPokemonGeneration, and: pokemonBaseDataModel.generation)
+        eggGroups.attributedText = setupAttributedLabels(with: mainPokemonEggGroups, and: pokemonBaseDataModel.eggGroups)
+        growthRate.attributedText = setupAttributedLabels(with: mainPokemonGrowthRate, and: pokemonBaseDataModel.growthRate)
+        habitat.attributedText = setupAttributedLabels(with: mainPokemonHabitat, and: pokemonBaseDataModel.habitat)
+        shape.attributedText = setupAttributedLabels(with: mainPokemonShape, and: pokemonBaseDataModel.shape)
     }
     
     private func setupAttributedLabels(with mainString: String, and stringForRange: String) -> NSAttributedString {